Separate APIs in the backend code
