+++ src/template/urls.py
@@ -16,12 +16,14 @@ Including another URLconf
 from django.conf.urls import include, url
 from django.contrib import admin
 
+from template.core import api as core_api
 from template.core import views as core_views
 
 urlpatterns = [
     url(r'^admin/', admin.site.urls),
     url('^', include('django.contrib.auth.urls')),
-    url(r'^me/password/$', core_views.change_password, name='change_password'),
+
+    url(r'^me/password/$', core_api.change_password, name='change_password'),
 
     url(r'^.*$', core_views.app_page, name='app_page'),
 ]
